requirements.txt
appdirs==1.4.4
attrs==21.2.0
azure-common==1.1.27
azure-core==1.17.0
azure-identity==1.6.0
azure-keyvault-secrets==4.3.0
black==21.7b0
bleach==4.0.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
docutils==0.17.1
essentials-configuration>=0.0.2
idna==3.2
importlib-metadata==4.6.3
iniconfig==1.1.1
isodate==0.6.0
isort==5.9.3
jeepney==0.7.1
keyring==23.0.1
msal==1.13.0
msal-extensions==0.3.0
msrest==0.6.21
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.0
pathspec==0.9.0
pkginfo==1.7.1
pluggy==0.13.1
portalocker==1.7.1
py==1.10.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
readme-renderer==29.0
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
toml==0.10.2
tomli==1.2.1
tqdm==4.62.0
twine==3.4.2
urllib3==1.26.6
webencodings==0.5.1
zipp==3.5.0
